rename hus.js to hustle.js, add legal stuff to footer
